refactor: split in groups
