+++ frontend/src/lib/constants/routes.constants.ts
@@ -14,6 +14,12 @@ export enum AppPath {
 
 // SvelteKit uses the group defined in src/routes/(app)/ as part of the routeId. It also prefixes it with /.
 export const ROUTE_ID_GROUP_APP = "/(app)";
+export const ROUTE_ID_GROUP_UNIVERSE = "/(u)";
+export const ROUTE_ID_GROUP_NNS = "/(nns)";
+export const ROUTE_ID_GROUP_LIST = "/(list)";
+export const ROUTE_ID_GROUP_DETAILS = "/(details)";
+
+export const ROUTE_ID_GROUPS = [ROUTE_ID_GROUP_APP, ROUTE_ID_GROUP_UNIVERSE, ROUTE_ID_GROUP_NNS, ROUTE_ID_GROUP_LIST, ROUTE_ID_GROUP_DETAILS];
 
 export const UNIVERSE_PARAM = "u";
