+++ frontend/src/routes/(app)/(nns)/canister/+layout.svelte
