+++ frontend/src/routes/(app)/(nns)/canister/+page.svelte
