+++ frontend/src/routes/(app)/(nns)/canisters/+layout.svelte
