+++ frontend/src/routes/(app)/(nns)/canisters/+page.svelte
