+++ frontend/src/routes/(app)/(nns)/launchpad/+layout.svelte
