+++ frontend/src/routes/(app)/(nns)/launchpad/+page.svelte
