+++ frontend/src/routes/(app)/(nns)/project/+layout.svelte
