+++ frontend/src/routes/(app)/(nns)/project/+page.svelte
