+++ frontend/src/routes/(app)/(u)/(details)/accounts/+layout.svelte
