+++ frontend/src/routes/(app)/(u)/(details)/neurons/+layout.svelte
