+++ frontend/src/routes/(app)/(u)/(details)/neurons/+page.svelte
