+++ frontend/src/routes/(app)/(u)/(details)/proposals/+layout.svelte
