+++ frontend/src/routes/(app)/(u)/(details)/proposals/+page.svelte
