+++ frontend/src/routes/(app)/(u)/(list)/neuron/+page.svelte
