+++ frontend/src/routes/(app)/(u)/(list)/neuron/+page.ts
@@ -1,7 +1,7 @@
 import { browser } from "$app/environment";
 import { NEURON_PARAM } from "$lib/constants/routes.constants";
 import type { LoadEvent } from "@sveltejs/kit";
-import type { PageLoad } from "./$types";
+import type { PageLoad } from "../../../../../.svelte-kit/types/src/routes";
 
 export const load: PageLoad = (
   $event: LoadEvent
