+++ frontend/src/routes/(app)/(u)/(list)/proposal/+layout.svelte
