+++ frontend/src/routes/(app)/(u)/(list)/proposal/+page.svelte
