+++ frontend/src/routes/(app)/(u)/(list)/wallet/+layout.svelte
