+++ frontend/src/routes/(app)/(u)/(list)/wallet/+page.svelte
