+++ frontend/src/tests/routes/app/neurons/page.spec.ts
@@ -3,7 +3,7 @@
  */
 
 import { authStore } from "$lib/stores/auth.store";
-import NeuronsPage from "$routes/(app)/neurons/+page.svelte";
+import NeuronsPage from "$routes/(app)/(u)/(details)/neurons/+page.svelte";
 import { render } from "@testing-library/svelte";
 import {
   authStoreMock,
